refactor: moved dynamics.math to sympy.math
The math sub-module is used by the kinematics sub-module later on. In
addition, it contains functionality that's not specific to dynamics.
