+++ src/ampform/sympy/__init__.py
