+++ src/ampform/sympy/math.py
