+++ tests/test_dynamics.py
@@ -1,15 +1,12 @@
 # pylint: disable=no-self-use, protected-access, too-many-arguments
 from typing import Tuple
 
-import numpy as np
-import pytest
 import sympy as sp
 from qrules import ParticleCollection
 from sympy import preorder_traversal
 
 from ampform.dynamics import (
     BlattWeisskopfSquared,
-    ComplexSqrt,
     EnergyDependentWidth,
     PhaseSpaceFactor,
     PhaseSpaceFactorAnalytic,
@@ -148,43 +145,3 @@ def round_nested(expression: sp.Expr, n_decimals: int) -> sp.Expr:
         if isinstance(node, sp.Pow) and node.args[1] == 1 / 2:
             expression = expression.subs(node, round(node.n(), n_decimals))
     return expression
-
-
-class TestComplexSqrt:
-    @pytest.mark.parametrize("real", [False, True])
-    def test_evaluate(self, real: bool):
-        x = sp.Symbol("x", real=real)
-        expr = ComplexSqrt(x).evaluate()
-        if real:
-            assert expr == sp.Piecewise(
-                (sp.I * sp.sqrt(-x), x < 0),
-                (sp.sqrt(x), True),
-            )
-        else:
-            assert expr == sp.sqrt(x)
-
-    def test_latex(self):
-        x = sp.Symbol("x")
-        expr = ComplexSqrt(x)
-        assert sp.latex(expr) == R"\sqrt[\mathrm{c}]{x}"
-
-    @pytest.mark.parametrize("real", [False, True])
-    @pytest.mark.parametrize("backend", ["math", "numpy"])
-    def test_lambdify(self, backend: str, real: bool):
-        x = sp.Symbol("x", real=real)
-        expression = ComplexSqrt(x)
-        lambdified = sp.lambdify(x, expression, backend)
-        assert lambdified(np.array(-1)) == 1j
-
-    @pytest.mark.parametrize(
-        ("input_value", "expected"),
-        [
-            (sp.Symbol("x", real=True), "ComplexSqrt(x)"),
-            (sp.Symbol("x"), "ComplexSqrt(x)"),
-            (+4, "2"),
-            (-4, "2*I"),
-        ],
-    )
-    def test_new(self, input_value, expected: str):
-        expr = ComplexSqrt(input_value)
-        assert str(expr) == expected
